requirement.txt
SpeechRecognition
pyaudio
google-generativeai
isodate
google-api-python-client
PyAudio